java/iot3mobility: move HazardType and HazardCategory to the roadobjects package
These objects are a representation of part of the underlying DENM, and not used anywhere in the its package. It seems more appropriate to locate them alongside the RoadHazard which is using them.
