+++ java/iot3/examples/src/main/java/com/orange/Iot3MobilityBootstrapExample.java
@@ -8,7 +8,7 @@ import com.orange.iot3mobility.IoT3MobilityCallback;
 import com.orange.iot3mobility.TrueTime;
 import com.orange.iot3mobility.Utils;
 import com.orange.iot3mobility.its.EtsiUtils;
-import com.orange.iot3mobility.its.HazardType;
+import com.orange.iot3mobility.roadobjects.HazardType;
 import com.orange.iot3mobility.its.StationType;
 import com.orange.iot3mobility.its.json.JsonValue;
 import com.orange.iot3mobility.its.json.Position;
