+++ java/iot3/mobility/src/main/java/com/orange/iot3mobility/roadobjects/HazardType.java
@@ -5,7 +5,7 @@
 
  @author Mathieu LEFEBVRE <mathieu1.lefebvre@orange.com>
  */
-package com.orange.iot3mobility.its;
+package com.orange.iot3mobility.roadobjects;
 
 public enum HazardType {
